+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryDark"
-    tools:context=".LoginActivity">
+    tools:context=".Activities.LoginActivity">
 
     <com.airbnb.lottie.LottieAnimationView
         android:id="@+id/imageView"
